# openapitools/openapi.yaml
openapi: 3.1.0
info:
    title: Train Travel API
    description: |
        API for finding and booking train trips across Europe.

        ## Run in Postman

        Experiment with this API in Postman, using our Postman Collection.

        [![Run In Postman](https://run.pstmn.io/button.svg =128pxx32px)](https://app.getpostman.com/run-collection/9265903-7a75a0d0-b108-4436-ba54-c6139698dc08?action=collection%2Ffork&source=rip_markdown&collection-url=entityId%3D9265903-7a75a0d0-b108-4436-ba54-c6139698dc08%26entityType%3Dcollection%26workspaceId%3Df507f69d-9564-419c-89a2-cb8e4c8c7b8f)
    version: 1.2.0
    contact:
        name: Train Support
        url: https://example.com/support
        email: support@example.com
    license:
        name: Creative Commons Attribution-NonCommercial-ShareAlike 4.0 International
        identifier: CC-BY-NC-SA-4.0
    x-feedbackLink:
        label: Submit Feedback
        url: https://github.com/bump-sh-examples/train-travel-api/issues/new
servers:
    - url: https://try.microcks.io/rest/Train+Travel+API/1.0.0
      description: Mock Server
      x-internal: false
    - url: https://api.example.com
      description: Production
      x-internal: false
...
security:
    - OAuth2:
        - read
...
paths:
    /stations:
        get:
            summary: Get a list of train stations
            description: Returns a paginated and searchable list of all train stations.
            operationId: get-stations
            tags:
                - Stations
            parameters:
                - $ref: '#/components/parameters/page'
                - $ref: '#/components/parameters/limit'
                - name: coordinates
                  in: query
                  description: >
                    The latitude and longitude of the user's location, to narrow down the search results to sites within a proximity of this location.

                  required: false
                  schema:
                    type: string
                  example: 52.5200,13.4050
                - name: search
                  in: query
                  description: >
                    A search term to filter the list of stations by name or address.

                  required: false
                  schema:
                    type: string
                - name: country
                  in: query
                  description: Filter stations by country code
                  required: false
                  schema:
                    type: string
                    format: iso-country-code
                  example: DE
            responses:
                '200':
                    description: OK
                    headers:
                        Cache-Control:
                            $ref: '#/components/headers/Cache-Control'
                        RateLimit:
                            $ref: '#/components/headers/RateLimit'
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/Wrapper-Collection'
                                    - properties:
                                        data:
                                            type: array
                                            items:
                                                $ref: '#/components/schemas/Station'
                                    - properties:
                                        links:
                                            allOf:
                                                - $ref: '#/components/schemas/Links-Self'
                                                - $ref: '#/components/schemas/Links-Pagination'
                            example:
                                data:
                                    - id: efdbb9d1-02c2-4bc3-afb7-6788d8782b1e
                                      name: Berlin Hauptbahnhof
                                      address: Invalidenstraße 10557 Berlin, Germany
                                      country_code: DE
                                      timezone: Europe/Berlin
                                    - id: b2e783e1-c824-4d63-b37a-d8d698862f1d
                                      name: Paris Gare du Nord
                                      address: 18 Rue de Dunkerque 75010 Paris, France
                                      country_code: FR
                                      timezone: Europe/Paris
                                links:
                                    self: https://api.example.com/stations&page=2
                                    next: https://api.example.com/stations?page=3
                                    prev: https://api.example.com/stations?page=1
                        application/xml:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/Wrapper-Collection'
                                    - properties:
                                        data:
                                            type: array
                                            xml:
                                                name: stations
                                                wrapped: true
                                            items:
                                                $ref: '#/components/schemas/Station'
                                    - properties:
                                        links:
                                            allOf:
                                                - $ref: '#/components/schemas/Links-Self'
                                                - $ref: '#/components/schemas/Links-Pagination'
                '400':
                    $ref: '#/components/responses/BadRequest'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '429':
                    $ref: '#/components/responses/TooManyRequests'
                '500':
                    $ref: '#/components/responses/InternalServerError'
    /trips:
        get:
            summary: Get available train trips
            description: >
                Returns a list of available train trips between the specified origin and destination stations on the given date, and allows for filtering by bicycle and dog allowances.

            operationId: get-trips
            tags:
                - Trips
            parameters:
                - $ref: '#/components/parameters/page'
                - $ref: '#/components/parameters/limit'
                - name: origin
                  in: query
                  description: The ID of the origin station
                  required: true
                  schema:
                    type: string
                    format: uuid
                  example: efdbb9d1-02c2-4bc3-afb7-6788d8782b1e
                - name: destination
                  in: query
                  description: The ID of the destination station
                  required: true
                  schema:
                    type: string
                    format: uuid
                  example: b2e783e1-c824-4d63-b37a-d8d698862f1d
                - name: date
                  in: query
                  description: The date and time of the trip in ISO 8601 format in origin station's timezone.
                  required: true
                  schema:
                    type: string
                    format: date-time
                  example: '2024-02-01T09:00:00Z'
                - name: bicycles
                  in: query
                  description: Only return trips where bicycles are known to be allowed
                  required: false
                  schema:
                    type: boolean
                    default: false
                - name: dogs
                  in: query
                  description: Only return trips where dogs are known to be allowed
                  required: false
                  schema:
                    type: boolean
                    default: false
            responses:
                '200':
                    description: A list of available train trips
                    headers:
                        Cache-Control:
                            $ref: '#/components/headers/Cache-Control'
                        RateLimit:
                            $ref: '#/components/headers/RateLimit'
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/Wrapper-Collection'
                                    - properties:
                                        data:
                                            type: array
                                            items:
                                                allOf:
                                                    - $ref: '#/components/schemas/Trip'
                                                    - $ref: '#/components/schemas/Links-Origin'
                                                    - $ref: '#/components/schemas/Links-Destination'
                                    - properties:
                                        links:
                                            allOf:
                                                - $ref: '#/components/schemas/Links-Self'
                                                - $ref: '#/components/schemas/Links-Pagination'
                            example:
                                data:
                                    - id: ea399ba1-6d95-433f-92d1-83f67b775594
                                      origin: efdbb9d1-02c2-4bc3-afb7-6788d8782b1e
                                      destination: b2e783e1-c824-4d63-b37a-d8d698862f1d
                                      departure_time: '2024-02-01T10:00:00Z'
                                      arrival_time: '2024-02-01T16:00:00Z'
                                      price: 50
                                      operator: Deutsche Bahn
                                      bicycles_allowed: true
                                      dogs_allowed: true
                                      links:
                                        self: https://api.example.com/trips/ea399ba1-6d95-433f-92d1-83f67b775594
                                        origin: https://api.example.com/stations/efdbb9d1-02c2-4bc3-afb7-6788d8782b1e
                                        destination: https://api.example.com/stations/b2e783e1-c824-4d63-b37a-d8d698862f1d
                                    - id: 4d67459c-af07-40bb-bb12-178dbb88e09f
                                      origin: b2e783e1-c824-4d63-b37a-d8d698862f1d
                                      destination: efdbb9d1-02c2-4bc3-afb7-6788d8782b1e
                                      departure_time: '2024-02-01T12:00:00Z'
                                      arrival_time: '2024-02-01T18:00:00Z'
                                      price: 50
                                      operator: SNCF
                                      bicycles_allowed: true
                                      dogs_allowed: true
                                      links:
                                        self: https://api.example.com/trips/4d67459c-af07-40bb-bb12-178dbb88e09f
                                        origin: https://api.example.com/stations/b2e783e1-c824-4d63-b37a-d8d698862f1d
                                        destination: https://api.example.com/stations/efdbb9d1-02c2-4bc3-afb7-6788d8782b1e
                                links:
                                    self: https://api.example.com/trips?origin=efdbb9d1-02c2-4bc3-afb7-6788d8782b1e&destination=b2e783e1-c824-4d63-b37a-d8d698862f1d&date=2024-02-01
                                    next: https://api.example.com/trips?origin=efdbb9d1-02c2-4bc3-afb7-6788d8782b1e&destination=b2e783e1-c824-4d63-b37a-d8d698862f1d&date=2024-02-01&page=2
                        application/xml:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/Wrapper-Collection'
                                    - properties:
                                        data:
                                            type: array
                                            xml:
                                                name: trips
                                                wrapped: true
                                            items:
                                                $ref: '#/components/schemas/Trip'
                                    - properties:
                                        links:
                                            allOf:
                                                - $ref: '#/components/schemas/Links-Self'
                                                - $ref: '#/components/schemas/Links-Pagination'
                '400':
                    $ref: '#/components/responses/BadRequest'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '429':
                    $ref: '#/components/responses/TooManyRequests'
                '500':
                    $ref: '#/components/responses/InternalServerError'
    /bookings:
        get:
            operationId: get-bookings
            summary: List existing bookings
            description: Returns a list of all trip bookings by the authenticated user.
            tags:
                - Bookings
            parameters:
                - $ref: '#/components/parameters/page'
                - $ref: '#/components/parameters/limit'
            responses:
                '200':
                    description: A list of bookings
                    headers:
                        Cache-Control:
                            $ref: '#/components/headers/Cache-Control'
                        RateLimit:
                            $ref: '#/components/headers/RateLimit'
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/Wrapper-Collection'
                                    - properties:
                                        data:
                                            type: array
                                            items:
                                                $ref: '#/components/schemas/Booking'
                                    - properties:
                                        links:
                                            allOf:
                                                - $ref: '#/components/schemas/Links-Self'
                                                - $ref: '#/components/schemas/Links-Pagination'
                            example:
                                data:
                                    - id: efdbb9d1-02c2-4bc3-afb7-6788d8782b1e
                                      trip_id: efdbb9d1-02c2-4bc3-afb7-6788d8782b1e
                                      passenger_name: John Doe
                                      has_bicycle: true
                                      has_dog: true
                                    - id: b2e783e1-c824-4d63-b37a-d8d698862f1d
                                      trip_id: b2e783e1-c824-4d63-b37a-d8d698862f1d
                                      passenger_name: Jane Smith
                                      has_bicycle: false
                                      has_dog: false
                                links:
                                    self: https://api.example.com/bookings
                                    next: https://api.example.com/bookings?page=2
                        application/xml:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/Wrapper-Collection'
                                    - properties:
                                        data:
                                            type: array
                                            xml:
                                                name: bookings
                                                wrapped: true
                                            items:
                                                $ref: '#/components/schemas/Booking'
                                    - properties:
                                        links:
                                            allOf:
                                                - $ref: '#/components/schemas/Links-Self'
                                                - $ref: '#/components/schemas/Links-Pagination'
                '400':
                    $ref: '#/components/responses/BadRequest'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '429':
                    $ref: '#/components/responses/TooManyRequests'
                '500':
                    $ref: '#/components/responses/InternalServerError'
        post:
            operationId: create-booking
            summary: Create a booking
            description: A booking is a temporary hold on a trip. It is not confirmed until the payment is processed.
            tags:
                - Bookings
            security:
                - OAuth2:
                    - write
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Booking'
                    application/xml:
                        schema:
                            $ref: '#/components/schemas/Booking'
            responses:
                '201':
                    description: Booking successful
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/Booking'
                                    - properties:
                                        links:
                                            $ref: '#/components/schemas/Links-Self'
                            example:
                                id: efdbb9d1-02c2-4bc3-afb7-6788d8782b1e
                                trip_id: efdbb9d1-02c2-4bc3-afb7-6788d8782b1e
                                passenger_name: John Doe
                                has_bicycle: true
                                has_dog: true
                                links:
                                    self: https://api.example.com/bookings/efdbb9d1-02c2-4bc3-afb7-6788d8782b1e
                        application/xml:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/Booking'
                                    - properties:
                                        links:
                                            $ref: '#/components/schemas/Links-Self'
                '400':
                    $ref: '#/components/responses/BadRequest'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '404':
                    $ref: '#/components/responses/NotFound'
                '409':
                    $ref: '#/components/responses/Conflict'
                '429':
                    $ref: '#/components/responses/TooManyRequests'
                '500':
                    $ref: '#/components/responses/InternalServerError'
    /bookings/{bookingId}:
        parameters:
            - name: bookingId
              in: path
              required: true
              description: The ID of the booking to retrieve.
              schema:
                type: string
                format: uuid
              example: 1725ff48-ab45-4bb5-9d02-88745177dedb
        get:
            summary: Get a booking
            description: Returns the details of a specific booking.
            operationId: get-booking
            tags:
                - Bookings
            responses:
                '200':
                    description: The booking details
                    headers:
                        Cache-Control:
                            $ref: '#/components/headers/Cache-Control'
                        RateLimit:
                            $ref: '#/components/headers/RateLimit'
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/Booking'
                                    - properties:
                                        links:
                                            $ref: '#/components/schemas/Links-Self'
                            example:
                                id: efdbb9d1-02c2-4bc3-afb7-6788d8782b1e
                                trip_id: efdbb9d1-02c2-4bc3-afb7-6788d8782b1e
                                passenger_name: John Doe
                                has_bicycle: true
                                has_dog: true
                                links:
                                    self: https://api.example.com/bookings/1725ff48-ab45-4bb5-9d02-88745177dedb
                        application/xml:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/Booking'
                                    - properties:
                                        links:
                                            $ref: '#/components/schemas/Links-Self'
                '400':
                    $ref: '#/components/responses/BadRequest'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '404':
                    $ref: '#/components/responses/NotFound'
                '429':
                    $ref: '#/components/responses/TooManyRequests'
                '500':
                    $ref: '#/components/responses/InternalServerError'
        delete:
            summary: Delete a booking
            description: Deletes a booking, cancelling the hold on the trip.
            operationId: delete-booking
            security:
                - OAuth2:
                    - write
            tags:
                - Bookings
            responses:
                '204':
                    description: Booking deleted
                '400':
                    $ref: '#/components/responses/BadRequest'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '404':
                    $ref: '#/components/responses/NotFound'
                '429':
                    $ref: '#/components/responses/TooManyRequests'
                '500':
                    $ref: '#/components/responses/InternalServerError'
...
components:
    parameters:
        page:
            name: page
            in: query
            description: The page number to return
            required: false
            schema:
                type: integer
                minimum: 1
                default: 1
            example: 1
        limit:
            name: limit
            in: query
            description: The number of items to return per page
            required: false
            schema:
                type: integer
                minimum: 1
                maximum: 100
                default: 10
            example: 10
    securitySchemes:
        OAuth2:
            type: oauth2
            description: OAuth 2.0 authorization code following RFC8725 best practices.
            flows:
                authorizationCode:
                    authorizationUrl: https://example.com/oauth/authorize
                    tokenUrl: https://example.com/oauth/token
                    scopes:
                        read: Read access
                        write: Write access
...
    headers:
        Cache-Control:
            description: "The Cache-Control header communicates directives for caching mechanisms in both requests and responses. \nIt is used to specify the caching directives in responses to prevent caches from storing sensitive information.\n"
            schema:
                type: string
                description: A comma-separated list of directives as defined in [RFC 9111](https://www.rfc-editor.org/rfc/rfc9111.html).
        RateLimit:
            description: |
                The RateLimit header communicates quota policies. It contains a `limit` to
                convey the expiring limit, `remaining` to convey the remaining quota units,
                and `reset` to convey the time window reset time.
            schema:
                type: string
...